parser now takes int max for all

+++ philo/src/parser.c
@@ -6,7 +6,7 @@
 /*   By: mevan-de <mevan-de@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2022/09/08 15:10:52 by mevan-de      #+#    #+#                 */
-/*   Updated: 2022/09/12 14:31:05 by mevan-de      ########   odam.nl         */
+/*   Updated: 2022/09/20 14:51:56 by mevan-de      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -37,9 +37,9 @@ bool	parse_input_to_info(int argc, char **argv, t_info *info)
 	if (!info)
 		return (false);
 	info->nr_philos = parse_arg_to_long(argv[1], INT_MAX);
-	info->time_to_die = parse_arg_to_long(argv[2], LONG_MAX);
-	info->time_to_eat = parse_arg_to_long(argv[3], LONG_MAX);
-	info->time_to_sleep = parse_arg_to_long(argv[4], LONG_MAX);
+	info->time_to_die = parse_arg_to_long(argv[2], INT_MAX);
+	info->time_to_eat = parse_arg_to_long(argv[3], INT_MAX);
+	info->time_to_sleep = parse_arg_to_long(argv[4], INT_MAX);
 	info->nr_times_to_eat = 0;
 	if (argc == 6)
 		info->nr_times_to_eat = parse_arg_to_long(argv[5], INT_MAX);
